hello-world-service/requirements.txt
astroid==2.13.2
black==22.12.0
click==8.1.3
colorama==0.4.6
dill==0.3.6
Flask==2.2.2
Flask-API==3.0.post1
gitdb==4.0.10
GitPython==3.1.30
importlib-metadata==6.0.0
isort==5.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.1
mccabe==0.7.0
mypy-extensions==0.4.3
pathspec==0.10.3
pip==22.3.1
platformdirs==2.6.2
py-healthcheck==1.10.1
PyJWT==2.6.0
pylint==2.15.10
setuptools==65.5.1
six==1.16.0
smmap==5.0.0
tomli==2.0.1
tomlkit==0.11.6
typed-ast==1.5.4
typing_extensions==4.4.0
Werkzeug==2.2.3
wrapt==1.14.1
zipp==3.11.0
